requirements.txt
# Core video processing
opencv-python>=4.8.0
moviepy>=1.0.3
pydub>=0.25.1

# ML and AI
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
scipy>=1.10.0
numpy>=1.24.0
pandas>=2.0.0

# Video analysis
scenedetect>=0.6.0

# Audio processing
librosa>=0.10.0

# Video downloading (for reference)
yt-dlp>=2023.7.6

# Web scraping and APIs
requests>=2.31.0
beautifulsoup4>=4.12.0

# YouTube API
google-api-python-client>=2.95.0
google-auth>=2.22.0

# Text-to-Speech
azure-cognitiveservices-speech>=1.30.0

# Database
sqlite3

# Web framework (for dashboard)
flask>=2.3.0

# Image processing
Pillow>=10.0.0

# Progress bars
tqdm>=4.65.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Anime database APIs
mal-api>=0.5.0

# Utilities
python-dateutil>=2.8.0